rules-commons : add javadoc, java7 specificities

+++ rules/rules-commons/src/main/java/org/opens/tanaguru/rules/elementchecker/contrast/helper/ContrastHelper.java
@@ -168,7 +168,7 @@ public final class ContrastHelper {
      * @return
      */
     private static double computeContrast(double lighter, double darker) {
-        return (Double.valueOf(((lighter + CONTRAST_FACTOR) / (darker + CONTRAST_FACTOR))));
+        return (((lighter + CONTRAST_FACTOR) / (darker + CONTRAST_FACTOR)));
     }
 
     /**
